style: 在PC上编译通过。

+++ src/include/lidar_slam/slam_param_def.h
@@ -8,6 +8,7 @@
 
 namespace lidar_slam{
 struct ExtrinsicParam{
+    // EIGEN_MAKE_ALIGNED_OPERATOR_NEW
     bool extrinsic_est_en;
     V3D extrinT; 
     M3D extrinR;
@@ -35,7 +36,7 @@ struct LidarPreprocParam{
     ///////////////////////////////////////////////////////
     float blind_distance;
     float max_distance;
-    std::vector<float> z_range = {-5.0, 20.0};
+    std::vector<double> z_range = {-5.0, 20.0};
     int keep_lidar_num_before_curr = 1;
     int point_filter_num = 2;
     int ring_filter_num = 1;
